corrected event placement
